pyproject.toml
[project]
name = "pokemon-finder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
]
